+++ demos/layout/menuIcons.tsx
@@ -1,32 +1,39 @@
+import {
+  FacebookOutlined,
+  StarOutlined,
+  TwitterOutlined,
+} from '@ant-design/icons';
 import { PageContainer, ProLayout } from '@ant-design/pro-components';
 
+/** 菜单 `icon` 支持 React 节点（含 @ant-design/icons）、或图片/SVG URL 字符串 */
 const Demo = () => (
-  <div
-    style={{
-      height: '100vh',
-    }}
-  >
+  <div style={{ height: '100vh' }}>
     <ProLayout
       location={{
         pathname: '/articles/new',
       }}
-      iconfontUrl="//at.alicdn.com/t/font_8d5l8fzk5b87iudi.js"
       route={{
         routes: [
           {
             path: '/home',
             name: '收藏',
-            icon: 'icon-shoucang1',
+            icon: <StarOutlined />,
           },
           {
             path: '/home/overview',
-            name: 'FaceBook',
-            icon: 'icon-facebook',
+            name: 'Facebook',
+            icon: <FacebookOutlined />,
           },
           {
             path: '/home/search',
             name: 'Twitter',
-            icon: 'icon-twitter',
+            icon: <TwitterOutlined />,
+          },
+          {
+            path: '/svg-asset',
+            name: 'SVG 资源',
+            icon:
+              'https://gw.alipayobjects.com/zos/antfincdn/UCSef1BATc/nice.svg',
           },
         ],
       }}