+++ src/form/components/ColorPicker/index.tsx
@@ -53,4 +53,6 @@ const ProFormColorPicker: React.ForwardRefRenderFunction<
   );
 };
 
-export default React.forwardRef(ProFormColorPicker);
+export default React.forwardRef(ProFormColorPicker) as React.ForwardRefExoticComponent<
+  ProFormColorPickerProps & React.RefAttributes<any>
+>;